+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Haskell List interfaces for std::vector
+   - fmap
+   - map
+   - filter
+   - foldl
+   - mempty and mappend
+   - liftA2
+   - abofa
 
 ## [0.5.0] - 2020-11-10
 ### Added